+++ Diskretka/MUL_ND_N.h
@@ -5,7 +5,7 @@ N MUL_ND_N(N a, int b)
 {
 	N c;
 	int temp = 0, l = 0, temp2 = 0;
-	for (size_t i = 0; i < a.len; i++)
+	for (int i = 0; i < a.len; i++)
 	{
 		c.n = (int*)realloc(c.n, (l + 1) * sizeof(int));
 		temp = a.n[i] * b;
